# [IMP] core: speedup new_test_user and authenticate in tests