The /test_http test suite makes a heavy usage of both `new_test_user` and `authenticate`. Those two functions hash the password using very slow and computing intensive hash algorithms, which can take up to 1 full second on a slow laptop, everytime.

Those `new_test_user` are never commited, and will never be. They solely exist within the transaction of the test and are discarded once the test finished running. Also the cleartext version of the password is visible in the source code so there is no point strengenthing those.

In this work we patch the `CryptContext` instance for `res.users` to inject a very weak but very fast hash algorithm in front of the secure ones. This makes creating and authenticating new users much faster, which makes running the tests much faster too.

Running the `/test_http` test suite takes half the time with this patch.

Without this patch:

    Executed in   48,89 secs    fish           external
       usr time   31,23 secs  600,00 micros   31,23 secs
       sys time    0,83 secs  312,00 micros    0,83 secs

With this patch:

    Executed in   26,78 secs    fish           external
       usr time    8,15 secs   21,08 millis    8,13 secs
       sys time    0,89 secs    4,38 millis    0,89 secs