@robodoo 2025-01-14
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/193643.png)](https://mergebot.odoo.com/odoo/odoo/pull/193643)