@Julien00859 2025-01-15
exists in 17.2 at odoo/odoo#152378